moves buffer_manager and render_manger into own mods
